修复当前版本下画偏移0.5格的问题

+++ core/src/main/java/org/geysermc/geyser/entity/type/Entity.java
@@ -144,7 +144,7 @@ public class Entity {
     protected void initializeMetadata() {
         dirtyMetadata.put(EntityData.SCALE, 1f);
         dirtyMetadata.put(EntityData.COLOR, (byte) 0);
-        dirtyMetadata.put(EntityData.MAX_AIR_SUPPLY, (short) getMaxAir());
+        dirtyMetadata.put(EntityData.MAX_AIR_SUPPLY, getMaxAir());
         setDimensions(Pose.STANDING);
         setFlag(EntityFlag.HAS_GRAVITY, true);
         setFlag(EntityFlag.HAS_COLLISION, true);
